[refactor] 拆分 images controllers 結構，統一變數命名
