+++ controllers/images/index.js
@@ -1,0 +1,5 @@
+const imageBase = require("./imageBase");
+
+module.exports = {
+  ...imageBase,
+};
